build with NDK r21, and try not to build fluidsynth executable.
Starting Android 11 it will become impossible to run standalone executable
on Android shell, so there is no point of building it (it was not useful on
earlier platforms either). Since it causes build failures (missing some
libc functions) it *has to* be disabled.

NDK is relocated from `~/android-sdk-[plat]/ndk-bundle` to
`~/Android/Sdk/ndk/[rev]` as well.